ajout carrousel + restructuration systeème de generator
