+++ index.html
@@ -8,11 +8,15 @@
 
     <link rel="stylesheet" href="css/style.css">
     <script src="https://kit.fontawesome.com/1a4ed4a819.js" crossorigin="anonymous"></script>
+    <script defer type="module" src="js/generatorCard.js"></script>
     <script defer type="module" src="js/api.js"></script>
     <script defer type="module" src="js/generator.js"></script>
-    <script defer type="module" src="js/script.js"></script>
 
+    <script defer type="module" src="js/script.js"></script>
     <script src="https://cdn.tailwindcss.com"></script>
+    <link defer href="https://cdn.jsdelivr.net/npm/flowbite@2.5.2/dist/flowbite.min.css" rel="stylesheet" />
+    <script defer  src="https://cdn.jsdelivr.net/npm/flowbite@2.5.2/dist/flowbite.min.js"></script>
+
 </head>
 
 <body>
@@ -137,6 +141,8 @@
 
 
     </div>
+
+
 </body>
 
 </html>
